feat: migrando projeto para typescript e conversão de classes para função
