+++ package.json
@@ -6,6 +6,8 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "@types/node": "^17.0.21",
+    "@types/react": "^17.0.39",
     "@unform/core": "^2.1.6",
     "@unform/web": "^2.1.6",
     "axios": "^0.21.1",
@@ -16,6 +18,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.2.1",
+    "typescript": "^4.5.5",
     "yup": "^0.32.9"
   },
   "scripts": {
@@ -44,6 +47,10 @@
     ]
   },
   "devDependencies": {
+    "@types/react-dom": "^17.0.11",
+    "@types/react-modal": "^3.13.1",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.23",
     "json-server": "^0.16.3"
   }
 }
