+++ src/pages/Dashboard/index.tsx
@@ -1,0 +1,123 @@
+import { useEffect, useState } from 'react';
+
+
+import Header from '../../components/Header';
+import api from '../../services/api';
+import Food from '../../components/Food';
+import ModalAddFood from '../../components/ModalAddFood';
+import ModalEditFood from '../../components/ModalEditFood';
+import { FoodsContainer } from './styles';
+
+interface FoodProps {
+  id: string;
+  name: string;
+  description: string;
+  image: string;
+  price: number;
+  available: boolean;
+}
+
+const Dashboard = () => {
+  const [foods, setFoods] = useState<FoodProps[]>([]);
+  const [editingFood, setEditingFood] = useState<FoodProps>({
+    id: "",
+    name: "",
+    description: "",
+    image: "",
+    price: 0,
+    available: false
+  });
+  const [modalOpen, setModalOpen] = useState(false);
+  const [editModalOpen, setEditModalOpen] = useState(false);
+
+  useEffect(() => {
+    async function fetchApi() {
+      const response = await api.get('/foods');
+      setFoods(response.data);
+    }
+    fetchApi()
+  },[])
+
+  async function handleAddFood(food: FoodProps) {
+    try {
+      const response = await api.post('/foods', {
+        ...food,
+        available: true,
+      });
+
+      setFoods([...foods, response.data]);
+    } catch (err) {
+      console.log(err);
+    }
+  }
+
+  async function handleUpdateFood(food: FoodProps) {
+    try {
+      const foodUpdated = await api.put(
+        `/foods/${editingFood.id}`,
+        { ...editingFood, ...food },
+      );
+
+      const foodsUpdated = foods.map(f =>
+        f.id !== foodUpdated.data.id ? f : foodUpdated.data,
+      );
+
+      setFoods(foodsUpdated);
+    } catch (err) {
+      console.log(err);
+    }
+  }
+
+  async function handleDeleteFood(id: string) {
+    await api.delete(`/foods/${id}`);
+
+    const foodsFiltered = foods.filter(food => food.id !== id);
+
+    setFoods(foodsFiltered);
+  }
+
+  function toggleModal()  {
+    setModalOpen(!modalOpen);
+  }
+
+  function toggleEditModal() {
+   setEditModalOpen(!editModalOpen);
+  }
+
+  function handleEditFood(food: FoodProps) {
+    setEditingFood(food)
+    setEditModalOpen(true)
+  }
+
+
+  return (
+    <>
+      <Header openModal={toggleModal} />
+      <ModalAddFood
+        isOpen={modalOpen}
+        setIsOpen={toggleModal}
+        handleAddFood={handleAddFood}
+      />
+      <ModalEditFood
+        isOpen={editModalOpen}
+        setIsOpen={toggleEditModal}
+        editingFood={editingFood}
+        handleUpdateFood={handleUpdateFood}
+      />
+
+      <FoodsContainer data-testid="foods-list">
+        {foods &&
+          foods.map(food => (
+            <Food
+              key={food.id}
+              food={food}
+              handleDelete={() => handleDeleteFood(food.id)}
+              handleEditFood={() => handleEditFood(food)}
+            />
+          ))}
+      </FoodsContainer>
+    </>
+  );
+};
+
+export default Dashboard;
